+++ sdk/merchantSession/log/log.js
@@ -5,26 +5,26 @@ const Resource = require('../../utils/resource.js').Resource
 class Log extends Resource {
 
     /**
-     * Check out our API Documentation at https://starkbank.com/docs/api#merchant-purchase
+     * Check out our API Documentation at https://starkbank.com/docs/api#merchant-session
      */
 
-    constructor({ created, type, errors, purchase, id }) {
+    constructor({ created, type, errors, session, id }) {
         super(id);
         this.created = check.datetime(created);
         this.type = type;
         this.errors = errors;
-        this.purchase = purchase;
+        this.session = session;
     }
 }
 
 exports.Log = Log;
-let resource = {'class': exports.Log, 'name': 'MerchantPurchaseLog'};
+let resource = {'class': exports.Log, 'name': 'MerchantSessionLog'};
 
 exports.get = async function (id, {user} = {}) {
     return rest.getId(resource, id, user);
 };
 
-exports.query = async function ({ limit, after, before, user} = {}) {
+exports.query = async function ({ limit, after, before, user } = {}) {
     let query = {
         limit: limit,
         after: after,
@@ -33,14 +33,14 @@ exports.query = async function ({ limit, after, before, user} = {}) {
     return rest.getList(resource, query, user);
 };
 
-exports.page = async function ({ cursor, limit, after, before, types, purchaseIds, user } = {}) {
+exports.page = async function ({ cursor, limit, after, before, types, sessionIds, user } = {}) {
     let query = {
         cursor: cursor,
         limit: limit,
         after: after,
         before: before,
         types: types,
-        purchaseIds: purchaseIds,
+        sessionIds: sessionIds,
     };
     return rest.getPage(resource, query, user);
 };